requirements.txt
cached_property==1.5.2
dash==2.14.2
dash_bootstrap_components==1.6.0
h5py==3.10.0
ipython==8.12.3
ipywidgets==8.1.3
matplotlib==3.8.0
matplotlib==3.8.3
numpy==1.25.2
opencv_python==4.9.0.80
pandas==2.2.2
Pillow==10.4.0
pycocotools==2.0.7
PyQt5==5.15.11
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.2.1
scipy==1.9.3
segment_anything==1.0
skimage==0.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.65.0
